+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ?
+### Added
+### Changed
+### Updated
+ - [Gigahorse](https://github.com/madMAx43v3r/chia-gigahorse/releases/tag/v2.5.3.giga36) to v2.5.3.giga36.  NOTE: Only for AMD64, so ARM64 uses an old version.
+
 ## [2.5.3] - 2025-03-28
 ### Added
 ### Changed